+++ proyecto-tad/src/main/java/upo/tad/g11/proyecto/tad/model/DAO/Connection.java
@@ -1,8 +1,3 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package upo.tad.g11.proyecto.tad.model.DAO;
 
 import com.mongodb.MongoClient;
