+++ proyecto-tad/src/main/java/upo/tad/g11/proyecto/tad/model/entity/Servicio.java
@@ -1,14 +1,8 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package upo.tad.g11.proyecto.tad.model.entity;
 
 import dev.morphia.annotations.Entity;
 import dev.morphia.annotations.Id;
 import java.io.Serializable;
-import org.bson.Document;
 import org.bson.types.ObjectId;
 
 /**
